Partially enable ReusableBuildersTest. Make List builder reusable.
